+++ src/components/board.js
@@ -1,0 +1,9 @@
+const getBoardTpl = () => {
+  return `
+    <div class="board container">
+      <div class="board__tasks"></div>
+    </div>
+  `;
+};
+
+export { getBoardTpl };
