+++ src/components/more-btn.js
@@ -1,0 +1,7 @@
+const getMoreBtnTpl = () => {
+  return `
+    <button class="load-more" type="button">load more</button>
+  `;
+};
+
+export { getMoreBtnTpl };
